requirements.txt
torch
torchvision
open_clip_torch
tqdm
numpy
pandas
pyyaml
matplotlib